custom_components/simplechatbot/demo/requirements.txt
gradio_simplechatbot